# frontend/packages/api-client/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AllocateStockRequest": {
        "properties": {
          "stockItemId": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Stockitemid",
            "type": "string"
          }
        },
        "required": [
          "stockItemId"
        ],
        "title": "AllocateStockRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Consistent error contract distinguishing validation/business/conflict/infrastructure outcomes.",
        "properties": {
          "detail": {
            "description": "Specific explanation for this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "title": {
            "description": "Short human-readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable machine-readable error category, e.g. 'not_found'.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Page_StockItemResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StockItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[StockItemResponse]",
        "type": "object"
      },
...
      "StockHierarchyLink": {
        "properties": {
          "displayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Displaynamechain",
            "type": "array"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "displayNameChain"
        ],
        "title": "StockHierarchyLink",
        "type": "object"
      },
      "StockItemCreateRequest": {
        "properties": {
          "entityId": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "productId": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Productid",
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/StockPropertiesTransport"
          },
          "type": {
            "enum": [
              "stock/airline/flight",
              "stock/accommodation/room-type",
              "stock/mobility/transfer",
              "stock/mobility/rail",
              "stock/mobility/coach",
              "stock/mobility/vehicle-rental",
              "stock/water-transport/day-boat",
              "stock/water-transport/cruise",
              "stock/experience/guided-tour",
              "stock/experience/activity",
              "stock/protection/travel"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "productId",
          "type",
          "properties"
        ],
        "title": "StockItemCreateRequest",
        "type": "object"
      },
      "StockItemResponse": {
        "properties": {
          "availabilityState": {
            "enum": [
              "available",
              "allocated",
              "shortfall",
              "withdrawn",
              "expired"
            ],
            "title": "Availabilitystate",
            "type": "string"
          },
          "availableQuantity": {
            "title": "Availablequantity",
            "type": "integer"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "StockItem",
            "default": "StockItem",
            "title": "Entitykind",
            "type": "string"
          },
          "productAncestors": {
            "items": {
              "$ref": "#/components/schemas/StockHierarchyLink"
            },
            "title": "Productancestors",
            "type": "array"
          },
          "productDisplayName": {
            "title": "Productdisplayname",
            "type": "string"
          },
          "productDisplayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Productdisplaynamechain",
            "type": "array"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productType": {
            "title": "Producttype",
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/StockProperties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "supplierDisplayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplierdisplayname"
          },
          "supplierOrganisationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplierorganisationid"
          },
          "supplierRole": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StockHierarchyLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "type",
          "schemaVersion",
          "properties",
          "productId",
          "productType",
          "productDisplayName",
          "productDisplayNameChain",
          "productAncestors",
          "supplierRole",
          "supplierOrganisationId",
          "supplierDisplayName",
          "availableQuantity",
          "availabilityState"
        ],
        "title": "StockItemResponse",
        "type": "object"
      },
      "StockItemUpdateRequest": {
        "properties": {
          "properties": {
            "$ref": "#/components/schemas/StockPropertiesTransport"
          },
          "type": {
            "enum": [
              "stock/airline/flight",
              "stock/accommodation/room-type",
              "stock/mobility/transfer",
              "stock/mobility/rail",
              "stock/mobility/coach",
              "stock/mobility/vehicle-rental",
              "stock/water-transport/day-boat",
              "stock/water-transport/cruise",
              "stock/experience/guided-tour",
              "stock/experience/activity",
              "stock/protection/travel"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "properties"
        ],
        "title": "StockItemUpdateRequest",
        "type": "object"
      },
      "StockProperties": {
        "additionalProperties": false,
        "properties": {
          "capacityQuantity": {
            "default": 1,
            "minimum": 0.0,
            "title": "Capacityquantity",
            "type": "integer"
          },
          "currencyCode": {
            "const": "EUR",
            "title": "Currencycode",
            "type": "string"
          },
          "inventoryStatusCode": {
            "default": "inventory/active",
            "enum": [
              "inventory/active",
              "inventory/withdrawn",
              "inventory/expired"
            ],
            "title": "Inventorystatuscode",
            "type": "string"
          },
          "remainingCapacity": {
            "default": 1,
            "minimum": 0.0,
            "title": "Remainingcapacity",
            "type": "integer"
          },
          "serviceDate": {
            "format": "date",
            "title": "Servicedate",
            "type": "string"
          },
          "unitPriceAmount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unitpriceamount",
            "type": "string"
          }
        },
        "required": [
          "serviceDate",
          "unitPriceAmount",
          "currencyCode"
        ],
        "title": "StockProperties",
        "type": "object"
      },
      "StockPropertiesTransport": {
        "additionalProperties": false,
        "properties": {
          "capacityQuantity": {
            "default": 1,
            "minimum": 0.0,
            "title": "Capacityquantity",
            "type": "integer"
          },
          "currencyCode": {
            "const": "EUR",
            "title": "Currencycode",
            "type": "string"
          },
          "inventoryStatusCode": {
            "default": "inventory/active",
            "enum": [
              "inventory/active",
              "inventory/withdrawn",
              "inventory/expired"
            ],
            "title": "Inventorystatuscode",
            "type": "string"
          },
          "remainingCapacity": {
            "default": 1,
            "minimum": 0.0,
            "title": "Remainingcapacity",
            "type": "integer"
          },
          "serviceDate": {
            "format": "date",
            "title": "Servicedate",
            "type": "string"
          },
          "unitPriceAmount": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Unitpriceamount"
          }
        },
        "required": [
          "serviceDate",
          "unitPriceAmount",
          "currencyCode"
        ],
        "title": "StockPropertiesTransport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shared capability-oriented API reused by Customer and Staff Interaction.",
    "title": "Christopher Columbus Travel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/orders/{order_id}/positions/{position_id}/stock": {
      "put": {
        "operationId": "allocateOrderPositionStock",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Allocate Stock",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{order_id}/positions/{position_id}/stock/{stock_item_id}": {
      "delete": {
        "operationId": "releaseOrderPositionStock",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stock_item_id",
            "required": true,
            "schema": {
              "title": "Stock Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Release Stock",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/stock-items": {
      "get": {
        "operationId": "listStockItems",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page's nextCursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's nextCursor.",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "serviceDateFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Servicedatefrom"
            }
          },
          {
            "in": "query",
            "name": "serviceDateTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Servicedateto"
            }
          },
          {
            "in": "query",
            "name": "availabilityState",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "available",
                    "allocated",
                    "shortfall",
                    "withdrawn",
                    "expired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Availabilitystate"
            }
          },
          {
            "in": "query",
            "name": "productType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Producttype"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StockItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Stock Items",
        "tags": [
          "stock-items"
        ]
      },
      "post": {
        "operationId": "createStockItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockItemCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Stock Item",
        "tags": [
          "stock-items"
        ]
      }
    },
    "/stock-items/{stock_item_id}": {
      "delete": {
        "operationId": "withdrawStockItem",
        "parameters": [
          {
            "in": "path",
            "name": "stock_item_id",
            "required": true,
            "schema": {
              "title": "Stock Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Withdraw Stock Item",
        "tags": [
          "stock-items"
        ]
      },
      "get": {
        "operationId": "getStockItem",
        "parameters": [
          {
            "in": "path",
            "name": "stock_item_id",
            "required": true,
            "schema": {
              "title": "Stock Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Stock Item",
        "tags": [
          "stock-items"
        ]
      },
      "put": {
        "operationId": "updateStockItem",
        "parameters": [
          {
            "in": "path",
            "name": "stock_item_id",
            "required": true,
            "schema": {
              "title": "Stock Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockItemUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Stock Item",
        "tags": [
          "stock-items"
        ]
      }
    }
  }
}